move foreachboundary into new file
