+++ src/CMakeLists.txt
@@ -211,6 +211,7 @@ add_library(parthenon
   utils/error_checking.hpp
   utils/error_checking.cpp
   utils/hash.hpp
+  utils/loop_utils.hpp
   utils/mpi_types.hpp
   utils/object_pool.hpp
   utils/partition_stl_containers.hpp
