+++ src/bvals/cc/boundary_communication.cpp
@@ -34,11 +34,13 @@
 #include "mesh/meshblock.hpp"
 #include "mesh/refinement_in_one.hpp"
 #include "utils/error_checking.hpp"
+#include "utils/loop_utils.hpp"
 
 namespace parthenon {
 namespace cell_centered_bvars {
 
-using namespace impl;
+using namespace loops;
+using namespace loops::shorthands;
 
 template <BoundaryType bound_type>
 TaskStatus SendBoundBufs(std::shared_ptr<MeshData<Real>> &md) {
