+++ src/bvals/cc/bvals_utils.hpp
@@ -32,103 +32,10 @@
 #include "mesh/mesh.hpp"
 #include "mesh/meshblock.hpp"
 #include "utils/error_checking.hpp"
+#include "utils/loop_utils.hpp"
 
 namespace parthenon {
 namespace cell_centered_bvars {
-
-namespace impl {
-
-using sp_mb_t = std::shared_ptr<MeshBlock>;
-using sp_mbd_t = std::shared_ptr<MeshBlockData<Real>>;
-using sp_cv_t = std::shared_ptr<CellVariable<Real>>;
-using nb_t = NeighborBlock;
-
-enum class LoopControl { cont, break_out };
-
-// Methods for wrapping a function that may or may not return a LoopControl
-// object. The first is enabled if the function returns a LoopControl and
-// just passes the returned object on. The second just calls the function,
-// ignores its return, and returns a LoopControl continue. These wrap the
-// function calls in the ForEachBoundary loop template to allow for breaking
-// out of the loop if desired
-template <class F, class... Args>
-inline auto func_caller(F func, Args &&...args) -> typename std::enable_if<
-    std::is_same<decltype(func(std::declval<Args>()...)), LoopControl>::value,
-    LoopControl>::type {
-  return func(std::forward<Args>(args)...);
-}
-
-template <class F, class... Args>
-inline auto func_caller(F func, Args &&...args) -> typename std::enable_if<
-    !std::is_same<decltype(func(std::declval<Args>()...)), LoopControl>::value,
-    LoopControl>::type {
-  func(std::forward<Args>(args)...);
-  return LoopControl::cont;
-}
-
-// Loop over boundaries (or shared geometric elements) for blocks contained
-// in MeshData, calling the passed function func for every boundary. Unifies
-// boundary looping that occurs in many places in the boundary communication
-// routines and allows for easy selection of a subset of the boundaries based
-// on the template parameter BoundaryType. [Really, this probably does not
-// need to be a template parameter, it could just be a function argument]
-template <BoundaryType bound = BoundaryType::any, class F>
-inline void ForEachBoundary(std::shared_ptr<MeshData<Real>> &md, F func) {
-  for (int block = 0; block < md->NumBlocks(); ++block) {
-    auto &rc = md->GetBlockData(block);
-    auto pmb = rc->GetBlockPointer();
-    for (auto &v : rc->GetCellVariableVector()) {
-      if (v->IsSet(Metadata::FillGhost)) {
-        for (int n = 0; n < pmb->pbval->nneighbor; ++n) {
-          auto &nb = pmb->pbval->neighbor[n];
-          if constexpr (bound == BoundaryType::local) {
-            if (nb.snb.rank != Globals::my_rank) continue;
-          } else if constexpr (bound == BoundaryType::nonlocal) {
-            if (nb.snb.rank == Globals::my_rank) continue;
-          } else if constexpr (bound == BoundaryType::flxcor_send) {
-            // Check if this boundary requires flux correction
-            if (nb.snb.level != pmb->loc.level - 1) continue;
-            // No flux correction required unless boundaries share a face
-            if (std::abs(nb.ni.ox1) + std::abs(nb.ni.ox2) + std::abs(nb.ni.ox3) != 1)
-              continue;
-          } else if constexpr (bound == BoundaryType::flxcor_recv) {
-            // Check if this boundary requires flux correction
-            if (nb.snb.level - 1 != pmb->loc.level) continue;
-            // No flux correction required unless boundaries share a face
-            if (std::abs(nb.ni.ox1) + std::abs(nb.ni.ox2) + std::abs(nb.ni.ox3) != 1)
-              continue;
-          } else if constexpr (bound == BoundaryType::restricted) {
-            // Check if restriction is required
-            if (nb.snb.level >= pmb->loc.level) continue;
-          }
-          if constexpr (bound == BoundaryType::restricted) {
-            IndexRange bni, bnj, bnk;
-            ComputeRestrictionBounds(bni, bnj, bnk, nb, pmb);
-            // This loop is only over {-1, 0, 1}^3 at most
-            for (int nk = bnk.s; nk <= bnk.e; ++nk) {
-              for (int nj = bnj.s; nj <= bnj.e; ++nj) {
-                for (int ni = bni.s; ni <= bni.e; ++ni) {
-                  int ntype = std::abs(ni) + std::abs(nj) + std::abs(nk);
-                  // skip myself or coarse levels; only the same level must be restricted
-                  if (ntype == 0 ||
-                      pmb->pbval->nblevel[nk + 1][nj + 1][ni + 1] != pmb->loc.level)
-                    continue;
-                  OffsetIndices offsets(nk, nj, ni);
-                  if (func_caller(func, pmb, rc, nb, v, offsets) ==
-                      LoopControl::break_out)
-                    return;
-                }
-              }
-            }
-          } else {
-            OffsetIndices junk;
-            if (func_caller(func, pmb, rc, nb, v, junk) == LoopControl::break_out) return;
-          }
-        }
-      }
-    }
-  }
-}
 
 inline std::tuple<int, int, std::string, int>
 SendKey(const std::shared_ptr<MeshBlock> &pmb, const NeighborBlock &nb,
@@ -165,6 +72,8 @@ ReceiveKey(const std::shared_ptr<MeshBlock> &pmb, const NeighborBlock &nb,
 template <BoundaryType bound_type, class COMM_MAP, class F>
 void InitializeBufferCache(std::shared_ptr<MeshData<Real>> &md, COMM_MAP *comm_map,
                            BvarsSubCache_t *pcache, F KeyFunc, bool initialize_flags) {
+  using namespace loops;
+  using namespace loops::shorthands;
   Mesh *pmesh = md->GetMeshPointer();
 
   using key_t = std::tuple<int, int, std::string, int>;
@@ -213,6 +122,8 @@ void InitializeBufferCache(std::shared_ptr<MeshData<Real>> &md, COMM_MAP *comm_m
 
 template <BoundaryType BOUND_TYPE, bool SENDER>
 inline auto CheckSendBufferCacheForRebuild(std::shared_ptr<MeshData<Real>> md) {
+  using namespace loops;
+  using namespace loops::shorthands;
   BvarsSubCache_t &cache = md->GetBvarsCache().GetSubCache(BOUND_TYPE, SENDER);
 
   bool rebuild = false;
@@ -243,6 +154,8 @@ inline auto CheckSendBufferCacheForRebuild(std::shared_ptr<MeshData<Real>> md) {
 
 template <BoundaryType BOUND_TYPE, bool SENDER>
 inline auto CheckReceiveBufferCacheForRebuild(std::shared_ptr<MeshData<Real>> md) {
+  using namespace loops;
+  using namespace loops::shorthands;
   BvarsSubCache_t &cache = md->GetBvarsCache().GetSubCache(BOUND_TYPE, SENDER);
 
   bool rebuild = false;
@@ -272,6 +185,8 @@ inline auto CheckReceiveBufferCacheForRebuild(std::shared_ptr<MeshData<Real>> md
 
 template <BoundaryType BOUND_TYPE, bool SENDER>
 inline auto CheckNoCommCacheForRebuild(std::shared_ptr<MeshData<Real>> md) {
+  using namespace loops;
+  using namespace loops::shorthands;
   BvarsSubCache_t &cache = md->GetBvarsCache().GetSubCache(BOUND_TYPE, SENDER);
 
   bool rebuild = false;
@@ -304,6 +219,8 @@ using F_BND_INFO = std::function<BndInfo(
 template <BoundaryType BOUND_TYPE, bool SENDER>
 inline void RebuildBufferCache(std::shared_ptr<MeshData<Real>> md, int nbound,
                                F_BND_INFO BndInfoCreator) {
+  using namespace loops;
+  using namespace loops::shorthands;
   BvarsSubCache_t &cache = md->GetBvarsCache().GetSubCache(BOUND_TYPE, SENDER);
   cache.bnd_info = BufferCache_t("bnd_info", nbound);
   cache.bnd_info_h = Kokkos::create_mirror_view(cache.bnd_info);
@@ -358,7 +275,6 @@ inline void RebuildBufferCache(std::shared_ptr<MeshData<Real>> md, int nbound,
   }
 }
 
-} // namespace impl
 } // namespace cell_centered_bvars
 } // namespace parthenon
